requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.7
djangorestframework==3.12.4
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
idna==3.3
pika==1.2.0
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sortedcontainers==2.3.0
sqlparse==0.4.2
urllib3==1.26.9
